+++ CHANGELOG.md
@@ -7,6 +7,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 
 ### Added (2026-03-25)
+- ter-011: AWS VPC with public/private subnets
+  - Script: scripts/bash/terraform_toolkit/networking/vpc-setup.sh
+  - Terraform: scripts/bash/terraform_toolkit/networking/terraform/
+  - Doc: docs/how-to/terraform-aws-vpc.md
 - lin-016: DNS server with BIND9
   - Script: scripts/bash/linux_toolkit/dns/bind9-server-setup.sh
   - Doc: docs/how-to/linux-dns-bind9.md